+++ routes/web.php
@@ -1,6 +1,8 @@
 <?php
 
 use App\Http\Controllers\ProfileController;
+use App\Http\Controllers\Settings;
+use App\Models\UserPermission;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -21,6 +23,10 @@ Route::middleware(['auth'])->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+});
+
+Route::middleware(['auth', 'permission:'. UserPermission::ADMIN_USER_PERMISSION])->group(function () {
+    Route::get('/settings', [Settings::class, 'index'])->name('settings');
 });
 
 require __DIR__.'/webPush.php';